refactor(mandataires): move some components around :
- rename components/mandataireComponents -> components/mandataires
- replace TableMesure with react-table (UI)
