+++ src/components/index.js
@@ -2,5 +2,7 @@ export { default as DummyTabs } from "./common/DummyTabs";
 export { default as Navigation } from "./communComponents/Navigation";
 export { default as Footer } from "./communComponents/Footer";
 export { default as Layout } from "./Layout";
-
+export { default as PromiseState } from "./common/PromiseState";
+export { default as Button } from "./common/Button";
+export { default as SearchButton } from "./communComponents/SearchButton";
 export { Pill, PillDispo } from "./common/Pill";
